add mongo collection entity and mapper

+++ src/main/java/yrs/emos/controller/UserController.java
@@ -27,7 +27,7 @@ public class UserController {
     @Autowired
     private RedisTemplate redisTemplate;
     @Value("${emos.jwt.cache-expire}")
-    private int cacheExpire;
+    private long cacheExpire;
 
     @PostMapping("/register")
     @ApiOperation("register")
